--- 03week/mastermindNew.js
@@ -1,100 +1,0 @@
-'use strict';
-
-const assert = require('assert');
-const readline = require('readline');
-const rl = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout
-});
-
-let board = [];
-let solution = '';
-const letters = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'];
-
-function printBoard() {
-  for (let i = 0; i < board.length; i++) {
-    console.log(board[i]);
-  }
-}
-
-function generateSolution() {
-  for (let i = 0; i < 4; i++) {
-    const randomIndex = getRandomInt(0, letters.length);
-    solution += letters[randomIndex];
-  }
-}
-
-function getRandomInt(min, max) {
-  return Math.floor(Math.random() * (max - min)) + min;
-}
-
-function generateHint(guess) {
-  // your code here
-  let solutionArr = solution.split("");
-  let guessArr = guess.split("");
-  let red = 0;
-  let white = 0;
-  for (var i = 0; i < 4; i++) {
-    var match = guessArr.indexOf(solutionArr[i]);
-    if (solutionArr[i] === guessArr[i]) {
-      red++;
-    }
-    if (match > -1) {
-      white++;
-      guessArr[match] = null;
-    }
-  }
-
-  white = white - red;
-
-  return (red + '-' + white);
-}
-
-
-
-function mastermind(guess) {
-  if (guess === solution) {
-    return ('You guessed it!');
-  }
-
-  generateHint(guess);
-}
-
-
-function getPrompt() {
-  rl.question('guess: ', (guess) => {
-    console.log(mastermind(guess));
-    printBoard();
-    getPrompt();
-  });
-}
-
-// Tests
-
-if (typeof describe === 'function') {
-  solution = 'abcd';
-  describe('#mastermind()', () => {
-    it('should register a guess and generate hints', () => {
-      mastermind('aabb');
-      assert.equal(board.length, 1);
-    });
-    it('should be able to detect a win', () => {
-      assert.equal(mastermind(solution), 'You guessed it!');
-    });
-  });
-
-  describe('#generateHint()', () => {
-    it('should generate hints', () => {
-      assert.equal(generateHint('abdc'), '2-2');
-    });
-    it('should generate hints if solution has duplicates', () => {
-      assert.equal(generateHint('aabb'), '1-1');
-    });
-
-  });
-
-} else {
-
-  generateSolution();
-  getPrompt();
-}
